Moved request modal to a separate page and added documentUploader component
