+++ src/styles/pages/admin.module.css
@@ -8,14 +8,14 @@
 
 .headingNav {
   display: flex;
-  justify-content: space-around;
+  width: 100%;
+  justify-content: space-between;
   align-items: center;
   background-color: #472d5b;
-  padding: 1%;
 }
 
 .dashboard {
-  padding-top: 1%;
+  padding: 1%;
   background-color: white;
 }
