+++ src/styles/pages/request.module.css
@@ -1,0 +1,7 @@
+.container {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    gap: 1rem;
+}
